Improve declaration file for "react-input-mask"
 - move declaration for previous major version into v1 folder
 - remove module declaration construction
