--- types/react-input-mask/v2/tslint.json
@@ -1,3 +1,0 @@
-{
-    "extends": "dtslint/dt.json"
-}
